requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alpha-vantage==3.0.0
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
certifi==2026.2.25
charset-normalizer==3.4.6
finnhub-python==2.4.27
frozenlist==1.3.3
greenlet==3.1.1
idna==3.10
importlib-metadata==6.7.0
multidict==6.0.5
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2026.1.post1
requests==2.31.0
six==1.17.0
sqlalchemy==2.0.48
typing-extensions==4.7.1
urllib3==2.0.7
yarl==1.9.4
zipp==3.15.0
